pyproject.toml
[tool.poetry]
name = "django-online-school"
version = "0.1.0"
description = ""
authors = ["Maria Shmatko <yaakt@mail.ru>"]
readme = "README.md"
packages = [{include = "django_online_school"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.1.3"
djangorestframework = "^3.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
